added a few new classes. Need to find solution to changing state
